fix checkstyle warnings in smartCache
