+++ ktapi-smartCache/src/main/java/de/_125m125/kt/ktapi/smartcache/ImmediateResult.java
@@ -1,4 +1,4 @@
-package de._125m125.kt.ktapi.smartCache;
+package de._125m125.kt.ktapi.smartcache;
 
 import de._125m125.kt.ktapi.core.results.Result;
