+++ ktapi-smartCache/src/test/java/de/_125m125/kt/ktapi/smartcache/caches/CacheDataTest.java
@@ -1,4 +1,4 @@
-package de._125m125.kt.ktapi.smartCache.caches;
+package de._125m125.kt.ktapi.smartcache.caches;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -11,7 +11,8 @@ import java.util.Optional;
 import org.junit.Before;
 import org.junit.Test;
 
-import de._125m125.kt.ktapi.smartCache.objects.TimestampedList;
+import de._125m125.kt.ktapi.smartcache.caches.CacheData;
+import de._125m125.kt.ktapi.smartcache.objects.TimestampedList;
 
 public class CacheDataTest {
     private CacheData<String> uut;
