+++ ktapi-smartCache/src/test/java/de/_125m125/kt/ktapi/smartcache/caches/PrependCacheDataTest.java
@@ -1,4 +1,4 @@
-package de._125m125.kt.ktapi.smartCache.caches;
+package de._125m125.kt.ktapi.smartcache.caches;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
@@ -9,7 +9,8 @@ import java.util.Optional;
 import org.junit.Before;
 import org.junit.Test;
 
-import de._125m125.kt.ktapi.smartCache.objects.TimestampedList;
+import de._125m125.kt.ktapi.smartcache.caches.PrependCacheData;
+import de._125m125.kt.ktapi.smartcache.objects.TimestampedList;
 
 public class PrependCacheDataTest {
