+++ ktapi-smartCache/src/test/java/de/_125m125/kt/ktapi/smartcache/caches/ReplaceOrPrependOrInvalidateOnEmptyCacheDataTest.java
@@ -1,4 +1,4 @@
-package de._125m125.kt.ktapi.smartCache.caches;
+package de._125m125.kt.ktapi.smartcache.caches;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -10,18 +10,18 @@ import java.util.Optional;
 import org.junit.Before;
 import org.junit.Test;
 
-import de._125m125.kt.ktapi.smartCache.objects.TimestampedList;
+import de._125m125.kt.ktapi.smartcache.objects.TimestampedList;
 
-public class RemoveOrReplaceOrPrependOrInvalidateOnEmptyCacheDataTest {
+public class ReplaceOrPrependOrInvalidateOnEmptyCacheDataTest {
 
-    RemoveOrReplaceOrPrependOrInvalidateOnEmptyCacheData<String> uut;
-    ClockExtension                                               clock;
+    ReplaceOrPrependOrInvalidateOnEmptyCacheData<String> uut;
+    ClockExtension                                       clock;
 
     @Before
     public void beforePrependCacheDataTest() throws Exception {
         this.clock = new ClockExtension();
-        this.uut = new RemoveOrReplaceOrPrependOrInvalidateOnEmptyCacheData<>(String.class,
-                String::hashCode, s -> s.equals("CB"), this.clock);
+        this.uut = new ReplaceOrPrependOrInvalidateOnEmptyCacheData<>(String.class,
+                String::hashCode, this.clock);
         this.clock.progress();
     }
 
@@ -30,32 +30,12 @@ public class RemoveOrReplaceOrPrependOrInvalidateOnEmptyCacheDataTest {
     }
 
     @Test
-    public void testRemovesMatches() {
-        fillUut();
-        uut.updateEntries(new String[] { "CB" });
-
-        final Optional<TimestampedList<String>> actual = this.uut.getAll();
-        assertTrue(actual.isPresent());
-        assertEquals(Arrays.asList("Aa", "Ca"), actual.get());
-    }
-
-    @Test
-    public void testRemovesAndReplacesMatches() {
-        fillUut();
-        uut.updateEntries(new String[] { "BB", "CB" });
-
-        final Optional<TimestampedList<String>> actual = this.uut.getAll();
-        assertTrue(actual.isPresent());
-        assertEquals(Arrays.asList("BB", "Ca"), actual.get());
-    }
-
-    @Test
     public void testUpdateEntriesInvalidatesIfEmptyList() throws Exception {
         this.uut.updateEntries(Arrays.asList("a"));
 
         final Optional<TimestampedList<String>> actual = this.uut.getAll();
         assertFalse(actual.isPresent());
-        assertEquals(2000l, this.uut.getLastInvalidationTime());
+        assertEquals(2000L, this.uut.getLastInvalidationTime());
     }
 
     @Test
@@ -64,7 +44,7 @@ public class RemoveOrReplaceOrPrependOrInvalidateOnEmptyCacheDataTest {
 
         final Optional<TimestampedList<String>> actual = this.uut.getAll();
         assertFalse(actual.isPresent());
-        assertEquals(2000l, this.uut.getLastInvalidationTime());
+        assertEquals(2000L, this.uut.getLastInvalidationTime());
     }
 
     @Test
@@ -88,4 +68,5 @@ public class RemoveOrReplaceOrPrependOrInvalidateOnEmptyCacheDataTest {
         assertTrue(actual.isPresent());
         assertEquals(Arrays.asList("Da", "Aa", "Ba", "Ca"), actual.get());
     }
+
 }
